修改Uow扩展函数
