+++ src/Riven.AspNetCore.Uow/AspNetCore/Mvc/Extensions/ActionDescriptorExtensions.cs
@@ -1,10 +1,12 @@
-using Microsoft.AspNetCore.Mvc.Abstractions;
-using Riven.Uow;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text;
+using Microsoft.AspNetCore.Mvc.Abstractions;
+using Riven.Uow;
+using Riven.Uow.Extensions;
+
 
 namespace Riven.AspNetCore.Mvc.Extensions
 {
@@ -14,33 +16,5 @@ namespace Riven.AspNetCore.Mvc.Extensions
         {
             return actionDescriptor.GetMethodInfo().GetUnitOfWorkAttributeOrNull();
         }
-
-
-        public static UnitOfWorkAttribute GetUnitOfWorkAttributeOrNull(this MethodInfo methodInfo)
-        {
-            var attrs = methodInfo.GetCustomAttributes(true)
-                            .OfType<UnitOfWorkAttribute>()
-                            .ToArray();
-
-            if (attrs.Length > 0)
-            {
-                return attrs[0];
-            }
-
-            attrs = methodInfo.DeclaringType.GetTypeInfo()
-                            .GetCustomAttributes(true)
-                            .OfType<UnitOfWorkAttribute>()
-                            .ToArray();
-
-            if (attrs.Length > 0)
-            {
-                return attrs[0];
-            }
-
-
-            return null;
-        }
-
-        
     }
 }
